video framerate in caps is a double, did the modifications to the related code. Use new .pitivi
git-svn-id: svn+ssh://svn.gnome.org/svn/pitivi/trunk@453 d3729300-e425-0410-8a4c-d956edccc248

+++ src/pitivi-projectsettings.h
@@ -110,7 +110,7 @@ GSList			*pitivi_projectsettings_list_make( );
 GstCaps			*pitivi_projectsettings_vcaps_create ( int width, int height, int framerate );
 GstCaps			*pitivi_projectsettings_acaps_create ( int rate, int channel, int depth );
 
-int			pitivi_projectsettings_get_videorate(PitiviProjectSettings *ps);
+gdouble			pitivi_projectsettings_get_videorate(PitiviProjectSettings *ps);
 int			pitivi_projectsettings_get_audiorate(PitiviProjectSettings *ps);
 int			pitivi_projectsettings_get_audiodepth(PitiviProjectSettings *ps);
 gboolean		pitivi_projectsettings_get_videosize (PitiviProjectSettings *ps,
